+++ CommonplaceBookApp/DocumentTableController.swift
@@ -6,7 +6,7 @@ import MiniMarkdown
 import UIKit
 
 /// Knows how to perform key actions with the document
-protocol DocumentTableControllerDelegate: AnyObject {
+public protocol DocumentTableControllerDelegate: AnyObject {
   /// Shows a TextEditViewController in the detail view.
   func showDetailViewController(_ detailViewController: UIViewController)
   /// Initiates studying.
@@ -15,6 +15,7 @@ protocol DocumentTableControllerDelegate: AnyObject {
   func documentTableDidDeleteDocument(with noteIdentifier: Note.Identifier)
   func showAlert(_ alertMessage: String)
   func showPage(with noteIdentifier: Note.Identifier)
+  func documentTableDidRequestReview()
 }
 
 /// Given a notebook, this class can manage a table that displays the hashtags and pages of that notebook.
@@ -22,10 +23,13 @@ public final class DocumentTableController: NSObject {
   /// Designated initializer.
   public init(
     tableView: UITableView,
-    notebook: NoteStorage
+    notebook: NoteStorage,
+    delegate: DocumentTableControllerDelegate
   ) {
     self.notebook = notebook
+    self.delegate = delegate
     tableView.register(DocumentTableViewCell.self, forCellReuseIdentifier: ReuseIdentifiers.documentCell)
+    tableView.register(ButtonCell.self, forCellReuseIdentifier: ReuseIdentifiers.review)
     let titleRenderer = RenderedMarkdown.makeTitleRenderer()
     self.dataSource = DataSource(tableView: tableView) { (tableView, indexPath, item) -> UITableViewCell? in
       switch item {
@@ -44,17 +48,47 @@ public final class DocumentTableController: NSObject {
         cell.backgroundColor = .secondarySystemBackground
         cell.textLabel?.text = hashtag
         return cell
+      case .review(let count):
+        // swiftlint:disable:next force_cast
+        let cell = tableView.dequeueReusableCell(withIdentifier: ReuseIdentifiers.review, for: indexPath) as! ButtonCell
+        cell.backgroundColor = .grailBackground
+        cell.button.setTitle("Review (\(count))", for: .normal)
+        cell.button.contentHorizontalAlignment = .left
+        cell.button.isEnabled = (count > 0) // swiftlint:disable:this empty_count
+        cell.button.accessibilityIdentifier = "study-button"
+        cell.tapHandler = { [weak delegate] in
+          delegate?.documentTableDidRequestReview()
+        }
+        return cell
       }
     }
     super.init()
     tableView.delegate = self
     tableView.refreshControl = refreshControl
+    let needsPerformUpdatesObserver = CFRunLoopObserverCreateWithHandler(nil, CFRunLoopActivity.beforeWaiting.rawValue, true, 0) { [weak self] _, _ in
+      self?.updateDataSourceIfNeeded()
+    }
+    CFRunLoopAddObserver(CFRunLoopGetMain(), needsPerformUpdatesObserver, CFRunLoopMode.commonModes)
     updateCardsPerDocument()
+  }
+
+  @objc private func didTapReview() {
+    delegate?.documentTableDidRequestReview()
   }
 
   public var challengeDueDate = Date() {
     didSet {
       updateCardsPerDocument()
+    }
+  }
+
+  private var needsPerformUpdates = false
+  private var isPerformingUpdates = false
+
+  private func updateDataSourceIfNeeded() {
+    if needsPerformUpdates, !isPerformingUpdates {
+      performUpdates(animated: true)
+      needsPerformUpdates = false
     }
   }
 
@@ -78,31 +112,38 @@ public final class DocumentTableController: NSObject {
   /// If non-nil, only pages with these identifiers will be shown.
   public var filteredPageIdentifiers: Set<Note.Identifier>? {
     didSet {
-      performUpdates(animated: true)
+      needsPerformUpdates = true
     }
   }
 
   /// If not empty, show a list of hashtags as a section in the table.
   public var hashtags: [String] = [] {
     didSet {
-      performUpdates(animated: true)
+      needsPerformUpdates = true
     }
   }
 
   /// If set, only show pages that contain this hashtag.
   public var filteredHashtag: String? {
     didSet {
-      performUpdates(animated: true)
+      needsPerformUpdates = true
+    }
+  }
+
+  /// The number of items that the learner can review.
+  public var reviewItemCount: Int = 0 {
+    didSet {
+      needsPerformUpdates = true
     }
   }
 
   /// Delegate.
-  internal weak var delegate: DocumentTableControllerDelegate?
+  private(set) weak var delegate: DocumentTableControllerDelegate?
 
   private let notebook: NoteStorage
   private var cardsPerDocument = [Note.Identifier: Int]() {
     didSet {
-      performUpdates(animated: true)
+      needsPerformUpdates = true
     }
   }
 
@@ -130,10 +171,15 @@ public final class DocumentTableController: NSObject {
       cardsPerDocument: cardsPerDocument,
       hashtags: hashtags,
       filteredHashtag: filteredHashtag,
-      filteredPageIdentifiers: filteredPageIdentifiers
+      filteredPageIdentifiers: filteredPageIdentifiers,
+      reviewItemCount: reviewItemCount
     )
     let reallyAnimate = animated && DocumentTableController.majorSnapshotDifferences(between: dataSource.snapshot(), and: snapshot)
-    dataSource.apply(snapshot, animatingDifferences: reallyAnimate)
+
+    isPerformingUpdates = true
+    dataSource.apply(snapshot, animatingDifferences: reallyAnimate) {
+      self.isPerformingUpdates = false
+    }
   }
 
   /// Compares lhs & rhs to see if the differences are worth animating.
@@ -144,12 +190,20 @@ public final class DocumentTableController: NSObject {
     // The only way to get through this loop and return false is if every item in the left hand
     // side and the right hand side, in order, have matching hashtags or page identifiers.
     // In that case, whatever difference that exists between the snapshots is "minor"
-    for (lhsItem, rhsItem) in zip(lhs.itemIdentifiers, rhs.itemIdentifiers) {
+    let itemsToCompare = zip(lhs.itemIdentifiers, rhs.itemIdentifiers)
+    for (lhsItem, rhsItem) in itemsToCompare {
       switch (lhsItem, rhsItem) {
       case (.page(let lhsPage), .page(let rhsPage)):
-        if lhsPage.pageKey != rhsPage.pageKey { return true }
+        if lhsPage.pageKey != rhsPage.pageKey {
+          return true
+        }
       case (.hashtag(let lhsHashtag), .hashtag(let rhsHashtag)):
-        if lhsHashtag != rhsHashtag { return true }
+        if lhsHashtag != rhsHashtag {
+          return true
+        }
+      case (.review, .review):
+        // These items do nothing to affect "major differences"
+        continue
       default:
         return true
       }
@@ -169,6 +223,8 @@ extension DocumentTableController: UITableViewDelegate {
       delegate?.showPage(with: viewProperties.pageKey)
     case .hashtag(let hashtag):
       delegate?.documentSearchResultsDidSelectHashtag(hashtag)
+    case .review:
+      delegate?.documentTableDidRequestReview()
     }
   }
 
@@ -198,7 +254,7 @@ extension DocumentTableController: UITableViewDelegate {
         studyAction.backgroundColor = UIColor.systemBlue
         actions.append(studyAction)
       }
-    case .hashtag:
+    case .hashtag, .review:
       // NOTHING
       break
     }
@@ -214,7 +270,7 @@ extension DocumentTableController: UITableViewDelegate {
     switch section {
     case .hashtags:
       label.text = "Hashtag"
-    case .documents:
+    case .documents, .review:
       return nil
     }
     return label
@@ -226,7 +282,7 @@ extension DocumentTableController: UITableViewDelegate {
     case .hashtags:
       let font = UIFont.preferredFont(forTextStyle: .subheadline)
       return font.lineHeight + 8
-    case .documents:
+    case .documents, .review:
       return 0
     }
   }
@@ -249,18 +305,23 @@ private extension DocumentTableController {
   enum DocumentSection {
     /// List of available hashtags
     case hashtags
+    /// Buttons to quiz/review on notebook contents
+    case review
     /// List of documents.
     case documents
   }
 
   enum Item: Hashable, CustomStringConvertible {
     case hashtag(String)
+    case review(Int)
     case page(ViewProperties)
 
     var description: String {
       switch self {
       case .hashtag(let hashtag):
         return hashtag
+      case .review:
+        return "Review"
       case .page(let viewProperties):
         return "Page \(viewProperties.pageKey)"
       }
@@ -280,6 +341,7 @@ private extension DocumentTableController {
   enum ReuseIdentifiers {
     static let documentCell = "DocumentCollectionViewCell"
     static let hashtag = "HashtagCell"
+    static let review = "ReviewCell"
   }
 
   static func cell(
@@ -299,15 +361,7 @@ private extension DocumentTableController {
     titleRenderer.markdown = viewProperties.noteProperties.title
     cell.titleLabel.attributedText = titleRenderer.attributedString
     cell.accessibilityLabel = viewProperties.noteProperties.title
-    var detailString = viewProperties.noteProperties.hashtags.joined(separator: ", ")
-    if viewProperties.cardCount > 0 {
-      if !detailString.isEmpty { detailString += ". " }
-      if viewProperties.cardCount == 1 {
-        detailString += "1 card."
-      } else {
-        detailString += "\(viewProperties.cardCount) cards."
-      }
-    }
+    let detailString = viewProperties.noteProperties.hashtags.joined(separator: ", ")
     cell.detailLabel.attributedText = NSAttributedString(
       string: detailString,
       attributes: [
@@ -316,6 +370,9 @@ private extension DocumentTableController {
       ]
     )
     cell.documentModifiedTimestamp = viewProperties.noteProperties.timestamp
+    if let font = titleRenderer.defaultAttributes[.font] as? UIFont {
+      cell.verticalPadding = max(20, font.lineHeight.roundedToScreenScale() * 1.5)
+    }
     return cell
   }
 
@@ -343,13 +400,16 @@ private extension DocumentTableController {
     cardsPerDocument: [Note.Identifier: Int],
     hashtags: [String],
     filteredHashtag: String?,
-    filteredPageIdentifiers: Set<Note.Identifier>?
+    filteredPageIdentifiers: Set<Note.Identifier>?,
+    reviewItemCount: Int
   ) -> Snapshot {
     var snapshot = Snapshot()
     if !hashtags.isEmpty {
       snapshot.appendSections([.hashtags])
       snapshot.appendItems(hashtags.map { Item.hashtag($0) })
     }
+    snapshot.appendSections([.review])
+    snapshot.appendItems([.review(reviewItemCount)])
     snapshot.appendSections([.documents])
 
     let propertiesFilteredByHashtag = notebook.allMetadata
@@ -375,5 +435,12 @@ private extension DocumentTableController {
     snapshot.appendItems(objects)
     DDLogDebug("Generating snapshot with \(objects.count) entries: \(objects)")
     return snapshot
+  }
+}
+
+private extension CGFloat {
+  func roundedToScreenScale(_ rule: FloatingPointRoundingRule = .toNearestOrAwayFromZero) -> CGFloat {
+    let scale: CGFloat = 1.0 / UIScreen.main.scale
+    return scale * (self / scale).rounded(rule)
   }
 }