+++ CommonplaceBookApp/StudyViewController.swift
@@ -29,6 +29,12 @@ public final class StudyViewController: UIViewController {
     super.init(nibName: nil, bundle: nil)
   }
 
+  private func finishStudySession() {
+    studySession.studySessionEndDate = Date()
+    delegate?.studyViewController(self, didFinishSession: studySession)
+    dismiss(animated: true, completion: nil)
+  }
+
   required init?(coder aDecoder: NSCoder) {
     fatalError("init(coder:) has not been implemented")
   }
@@ -47,27 +53,100 @@ public final class StudyViewController: UIViewController {
   /// Just holds the "correct" versus "incorrect" color whilst swiping
   private lazy var colorWashView: UIView = {
     let view = UIView(frame: .zero)
-    view.backgroundColor = .clear
     return view
   }()
 
-  /// Holds the description of what the swipe action is doing.
-  /// `currentDynamicItem` adjusts its alpha based on how far the swipe is proceeding
-  private lazy var swipeDescriptionLabel: UILabel = {
-    let label = UILabel(frame: .zero)
-    label.font = UIFont.preferredFont(forTextStyle: .headline)
-    label.textColor = .label
-    return label
+  private lazy var blurView: UIVisualEffectView = {
+    let view = UIVisualEffectView(effect: UIBlurEffect(style: .systemUltraThinMaterial))
+    return view
   }()
 
-  /// The current text we are showing in the swipe description.
-  /// Memoized here to minimize layout calculations while panning.
-  private var swipeDescriptionText: String? {
-    didSet {
-      if oldValue != swipeDescriptionText {
-        swipeDescriptionLabel.text = swipeDescriptionText
-        swipeDescriptionLabel.sizeToFit()
+  private struct Swipe: Identifiable {
+    var id: CGVector.Direction { classifier.direction }
+    let classifier: LinearPanTranslationClassifier
+    let message: String
+    let color: UIColor
+    let requiresVisibleAnswer: Bool
+    let correct: Bool?
+    let shouldDismiss: Bool
+    let snapPoint: (UIView) -> CGPoint
+
+    func makeLabel() -> UILabel {
+      let label = UILabel(frame: .zero)
+      label.font = UIFont.preferredFont(forTextStyle: .headline)
+      label.textColor = .label
+      label.text = message
+      label.sizeToFit()
+      return label
+    }
+  }
+
+  private let swipeClassifiers: [Swipe] = [
+    Swipe(
+      classifier: LinearPanTranslationClassifier(direction: .down),
+      message: "Dismiss",
+      color: .clear,
+      requiresVisibleAnswer: false,
+      correct: nil,
+      shouldDismiss: true,
+      snapPoint: { CGPoint(x: $0.center.x, y: $0.center.y + $0.frame.height) }
+    ),
+    Swipe(
+      classifier: LinearPanTranslationClassifier(direction: .right),
+      message: "I got it right",
+      color: .systemGreen,
+      requiresVisibleAnswer: true,
+      correct: true,
+      shouldDismiss: false,
+      snapPoint: { CGPoint(x: $0.center.x + $0.frame.width, y: $0.center.y) }
+    ),
+    Swipe(
+      classifier: LinearPanTranslationClassifier(direction: .left),
+      message: "Need to review",
+      color: .systemRed,
+      requiresVisibleAnswer: true,
+      correct: false,
+      shouldDismiss: false,
+      snapPoint: { CGPoint(x: $0.center.x - $0.frame.width, y: $0.center.y) }
+    ),
+  ]
+
+  /// Given a pan gesture in the view and whether or not the answer is currently visible, returns the best Swipe
+  /// corresponding to that gesture and a mesure of how strongly we match.
+  private func bestSwipe(for vector: CGVector, answerVisible: Bool) -> (Swipe, CGFloat)? {
+    swipeClassifiers
+      .compactMap { swipe -> (Swipe, CGFloat)? in
+        if swipe.requiresVisibleAnswer, !answerVisible {
+          return nil
+        }
+        return (swipe, swipe.classifier.matchStrength(vector: vector))
       }
+      .reduce(nil) { (priorResult, tuple) -> (Swipe, CGFloat)? in
+        guard let result = priorResult else {
+          return tuple
+        }
+        if tuple.1 > result.1 {
+          return tuple
+        } else {
+          return result
+        }
+      }
+  }
+
+  /// Mapping of possible swipe gestures with labels that tell the person we are going to match that swipe.
+  private var statusMessageLabels: [Swipe.ID: UILabel] = [:]
+
+  /// Sets the alpha of a particular swipe label to `alpha` and all other labels to zero.
+  private func setAlpha(_ alpha: CGFloat, for swipe: Swipe) {
+    for (swipeID, label) in statusMessageLabels {
+      label.alpha = (swipeID == swipe.id) ? alpha : 0
+    }
+  }
+
+  /// Sets the alpha for all swipe messages to 0.
+  private func hideAllSwipeMessages() {
+    for (_, label) in statusMessageLabels {
+      label.alpha = 0
     }
   }
 
@@ -84,21 +163,31 @@ public final class StudyViewController: UIViewController {
       oldValue?.alpha = 0
       oldValue?.removeFromSuperview()
       cardSnapBehavior.map { animator.removeBehavior($0) }
-      currentDynamicItem = currentCardView.map { ColorTranslatingDynamicItem(view: $0, colorWashView: colorWashView, swipeDescriptionLabel: swipeDescriptionLabel, centerX: view.center.x) }
+      currentDynamicItem = currentCardView.map { ColorTranslatingDynamicItem(view: $0, origin: view.center, viewController: self) }
+      attachPanGestureRecognizer()
       UIView.animate(withDuration: 0.2, animations: {
         self.colorWashView.backgroundColor = self.colorWashView.backgroundColor?.withAlphaComponent(0)
+        self.hideAllSwipeMessages()
         self.currentCardView?.alpha = 1
-      }) { _ in
+      }) { finished in
+        DDLogInfo("Animation finished = \(finished)")
         if let current = self.currentCardView {
           current.becomeFirstResponder()
           current.accessibilityIdentifier = "current-card"
-          // TODO: Refactor so I don't have to force unwrap
-          let snapBehavior = UISnapBehavior(item: self.currentDynamicItem!, snapTo: self.view.center)
-          self.animator.addBehavior(snapBehavior)
-          self.cardSnapBehavior = snapBehavior
+          self.attachSnapBehavior()
         }
       }
     }
+  }
+
+  private func attachSnapBehavior() {
+    // TODO: Refactor so I don't have to force unwrap
+    guard let dynamicItem = currentDynamicItem, dynamicItem.bounds.size != .zero else {
+      return
+    }
+    let snapBehavior = UISnapBehavior(item: dynamicItem, snapTo: view.center)
+    animator.addBehavior(snapBehavior)
+    cardSnapBehavior = snapBehavior
   }
 
   private func attachPanGestureRecognizer() {
@@ -123,62 +212,77 @@ public final class StudyViewController: UIViewController {
   public override func viewDidLoad() {
     super.viewDidLoad()
     view.addSubview(colorWashView)
-    view.addSubview(swipeDescriptionLabel)
+    view.addSubview(blurView)
+    makeStatusLabels()
     view.addSubview(doneImageView)
     view.addSubview(progressView)
     colorWashView.snp.makeConstraints { make in
       make.edges.equalToSuperview()
     }
+    blurView.snp.makeConstraints { make in
+      make.edges.equalToSuperview()
+    }
     doneImageView.snp.makeConstraints { make in
-      make.right.top.equalTo(view.safeAreaLayoutGuide).inset(16)
+      make.right.bottom.equalTo(view.safeAreaLayoutGuide).inset(16)
     }
     progressView.snp.makeConstraints { make in
       make.left.equalTo(view.safeAreaLayoutGuide).inset(16)
       make.centerY.equalTo(doneImageView.snp.centerY)
       make.right.equalTo(doneImageView.snp.left).offset(-8)
     }
-    swipeDescriptionLabel.snp.makeConstraints { make in
-      make.centerX.equalToSuperview()
-      make.top.equalTo(doneImageView.snp.bottom).offset(16)
+    for (_, label) in statusMessageLabels {
+      label.snp.makeConstraints { make in
+        make.centerX.equalToSuperview()
+        make.top.equalTo(doneImageView).offset(-16)
+      }
     }
     studySession.studySessionStartDate = Date()
-    view.backgroundColor = .grailGroupedBackground
     configureUI(animated: false, completion: nil)
     // Assumes we're presented in a navigation controller
     navigationController?.presentationController?.delegate = self
   }
 
-  public override func viewDidLayoutSubviews() {
-    super.viewDidLayoutSubviews()
-    currentDynamicItem?.centerX = view.center.x
+  private func makeStatusLabels() {
+    for swipe in swipeClassifiers {
+      let label = swipe.makeLabel()
+      statusMessageLabels[swipe.id] = label
+      view.addSubview(label)
+    }
+  }
+
+  public override func viewDidAppear(_ animated: Bool) {
+    super.viewDidAppear(animated)
+    attachSnapBehavior()
   }
 
   @objc private func didPan(sender: UIPanGestureRecognizer) {
-    guard let currentCard = currentDynamicItem, let snap = cardSnapBehavior else { return }
+    guard
+      let currentCard = currentDynamicItem,
+      let isAnswerVisible = currentCardView?.isAnswerVisible,
+      let snap = cardSnapBehavior
+    else {
+      return
+    }
     let translation = sender.translation(in: currentCard.view)
     switch sender.state {
     case .began:
       cardSnapBehavior.map { animator.removeBehavior($0) }
     case .changed:
       currentCard.center = view.center + translation
-      if translation.x > 0 {
-        swipeDescriptionLabel.text = "I got it right"
-      } else {
-        swipeDescriptionLabel.text = "Need to review"
-      }
     case .ended:
       var correct: Bool?
-      if translation.x > 100.0 {
-        snap.snapPoint = CGPoint(x: view.center.x + view.frame.width, y: view.center.y)
-        correct = true
-      } else if translation.x < -100.0 {
-        snap.snapPoint = CGPoint(x: view.center.x - view.frame.width, y: view.center.y)
-        correct = false
-      } else {
-        snap.snapPoint = view.center
+      var shouldDismiss = false
+      if let (swipe, strength) = bestSwipe(for: CGVector(destination: translation), answerVisible: isAnswerVisible), strength >= 1 {
+        correct = swipe.correct
+        shouldDismiss = swipe.shouldDismiss
+        snap.snapPoint = swipe.snapPoint(view)
       }
       cardSnapBehavior.map { animator.addBehavior($0) }
-      if let correct = correct {
+      if shouldDismiss {
+        DispatchQueue.main.asyncAfter(deadline: .now() + 0.3) {
+          self.finishStudySession()
+        }
+      } else if let correct = correct {
         DispatchQueue.main.asyncAfter(deadline: .now() + 0.3) {
           // TODO: Get rid of force unwrap
           self.userDidRespond(correct: correct)
@@ -251,13 +355,11 @@ public final class StudyViewController: UIViewController {
         )
       )
       challengeView.delegate = self
-      challengeView.backgroundColor = .grailSecondaryGroupedBackground
       view.addSubview(challengeView)
       challengeView.snp.makeConstraints { make in
         make.left.right.equalTo(self.view.readableContentGuide)
         make.centerY.equalToSuperview()
       }
-      challengeView.layer.cornerRadius = 8
       completion(challengeView)
     } catch {
       DDLogError("Unexpected error generating challenge view: \(error)")
@@ -276,56 +378,60 @@ extension StudyViewController: UIAdaptivePresentationControllerDelegate {
   }
 }
 
-/// Changes the color wash of `colorWashView` as the position of `view` changes.
-private final class ColorTranslatingDynamicItem: NSObject, UIDynamicItem {
-  init(view: UIView, colorWashView: UIView, swipeDescriptionLabel: UILabel, centerX: CGFloat) {
-    self.view = view
-    self.colorWashView = colorWashView
-    self.swipeDescriptionLabel = swipeDescriptionLabel
-    self.centerX = centerX
-  }
-
-  let view: UIView
-  private let colorWashView: UIView
-  private let swipeDescriptionLabel: UILabel
-
-  var centerX: CGFloat {
-    didSet {
-      configureUI()
+private extension StudyViewController {
+  /// Changes the color wash of `colorWashView` as the position of `view` changes.
+  final class ColorTranslatingDynamicItem: NSObject, UIDynamicItem {
+    init(view: UIView, origin: CGPoint, viewController: StudyViewController) {
+      self.view = view
+      self.origin = origin
+      self.viewController = viewController
     }
-  }
 
-  var center: CGPoint {
-    get { view.center }
-    set {
-      view.center = newValue
-      configureUI()
+    let view: UIView
+    weak var viewController: StudyViewController?
+    var shouldChangeColor = false
+
+    var origin: CGPoint {
+      didSet {
+        configureUI()
+      }
     }
-  }
 
-  var bounds: CGRect {
-    get { view.bounds }
-    set { view.bounds = newValue }
-  }
+    var center: CGPoint {
+      get { view.center }
+      set {
+        view.center = newValue
+        configureUI()
+      }
+    }
 
-  var transform: CGAffineTransform {
-    get { view.transform }
-    set { view.transform = newValue }
-  }
+    var bounds: CGRect {
+      get { view.bounds }
+      set { view.bounds = newValue }
+    }
 
-  private func configureUI() {
-    let deltaX = center.x - centerX
-    let colorWash = deltaX < 0 ? UIColor.systemRed : UIColor.systemGreen
-    let alpha = min(abs(deltaX) / 100.0, 1.0)
-    swipeDescriptionLabel.alpha = alpha
-    let intensity = alpha * 0.4
-    colorWashView.backgroundColor = colorWash.withAlphaComponent(intensity)
+    var transform: CGAffineTransform {
+      get { view.transform }
+      set { view.transform = newValue }
+    }
+
+    private func configureUI() {
+      let vector = CGVector(origin: origin, destination: center)
+      if let (swipe, alpha) = viewController?.bestSwipe(for: vector, answerVisible: shouldChangeColor) {
+        let intensity = alpha * 0.4
+        viewController?.setAlpha(alpha, for: swipe)
+        viewController?.colorWashView.backgroundColor = swipe.color.withAlphaComponent(intensity)
+      } else {
+        viewController?.hideAllSwipeMessages()
+        viewController?.colorWashView.backgroundColor = .clear
+      }
+    }
   }
 }
 
 extension StudyViewController: ChallengeViewDelegate {
   public func challengeViewDidRevealAnswer(_ challengeView: ChallengeView) {
-    attachPanGestureRecognizer()
+    currentDynamicItem?.shouldChangeColor = true
   }
 }
 
@@ -336,5 +442,11 @@ extension CGPoint {
 
   static func += (lhs: inout CGPoint, rhs: CGPoint) {
     lhs = CGPoint(x: lhs.x + rhs.x, y: lhs.y + rhs.y)
+  }
+}
+
+extension CGFloat {
+  func plusOrMinus(_ delta: CGFloat) -> Range<CGFloat> {
+    return (self - delta) ..< (self + delta)
   }
 }